Move non-commands to subdir
